refactor: utils.py -> utils module
